Redid js code
